rpi_software/requirements.txt
paho_mqtt
picamera2
